add parallax to menu, part 4

+++ src/ram.h
@@ -10,7 +10,7 @@ __attribute__((section(".aligned.0"),retain)) struct OAM_BUF {
 unsigned char gamestate=0xff;
 
 unsigned char song;
-unsigned short interrupt_scroll;
+unsigned long interrupt_scroll;
